+++ src/errorHandler.ts
@@ -50,11 +50,13 @@ import {
   AccessTokenExpiredError,
   AccessTokenUnknownError,
   DuplicateRequestError,
+  IDFormatError,
 } from './errors';
 
 export const ErrorClassDefinitionsMap = {
   1: ServerError,
   10: NoPermissionError,
+  12: IDFormatError,
   13: EmptyBodyError,
   14: MissingRequiredFieldsError,
   15: FieldFormatError,