+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.1.0]
 
-- Added findByName method to Data Transitions service.
+- Added `findByName` method to Data Transitions service.
+- Added `findById`, `findByName` and `findFirst` helpers to services having a generic `find` function.
 
 ## [3.0.2]